+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.4
 * feature: `props.sortConfig.stringLocaleCompare` 控制是否使用 Array.prototype.localeCompare 排序字符串。默认true。
 * optimize: 优化虚拟滚动 pageSize(减去表头高度)
+* optimize: `StkTableColumn` 中的 `width`,`minWidth`,`maxWidth` 支持配置数字。
 * fix: `PageUp`/`PageDown` 按键翻页会大于一页的情况。
 * fix: `props.stripe` 时，虚拟滚动导致斑马纹错位问题。
 * fix: `props.theme` 改变时，高亮颜色有误的问题。